- Moves 'fitness' story template to 'no-days-off'.
