+++ packages/stickers/src/no-days-off/cta.js
