+++ packages/stickers/src/no-days-off/index.js
