+++ packages/templates/src/raw/no-days-off/template.json
